+++ esphome/components/climate_ir/climate_ir.h
@@ -1,0 +1,53 @@
+#pragma once
+
+#include "esphome/components/climate/climate.h"
+#include "esphome/components/remote_base/remote_base.h"
+#include "esphome/components/remote_transmitter/remote_transmitter.h"
+#include "esphome/components/sensor/sensor.h"
+
+namespace esphome {
+namespace climate {
+
+/* A base for climate which works by sending (and receiving) IR codes
+
+    To send IR codes implement
+      void ClimateIR::transmit_state_()
+
+    Likewise to decode a IR into the AC state, implement
+      bool RemoteReceiverListener::on_receive(remote_base::RemoteReceiveData data) and return true
+*/
+class ClimateIR : public climate::Climate, public Component, public remote_base::RemoteReceiverListener {
+ public:
+  ClimateIR(float minimum_temperature, float maximum_temperature, float temperature_step = 1.0f) {
+    this->minimum_temperature_ = minimum_temperature;
+    this->maximum_temperature_ = maximum_temperature;
+    this->temperature_step_ = temperature_step;
+  }
+
+  void setup() override;
+  void set_transmitter(remote_transmitter::RemoteTransmitterComponent *transmitter) {
+    this->transmitter_ = transmitter;
+  }
+  void set_supports_cool(bool supports_cool) { this->supports_cool_ = supports_cool; }
+  void set_supports_heat(bool supports_heat) { this->supports_heat_ = supports_heat; }
+  void set_sensor(sensor::Sensor *sensor) { this->sensor_ = sensor; }
+
+ protected:
+  float minimum_temperature_, maximum_temperature_, temperature_step_;
+
+  /// Override control to change settings of the climate device.
+  void control(const climate::ClimateCall &call) override;
+  /// Return the traits of this controller.
+  climate::ClimateTraits traits() override;
+
+  /// Transmit via IR the state of this climate controller.
+  virtual void transmit_state() {}
+
+  bool supports_cool_{true};
+  bool supports_heat_{true};
+
+  remote_transmitter::RemoteTransmitterComponent *transmitter_;
+  sensor::Sensor *sensor_{nullptr};
+};
+}  // namespace climate
+}  // namespace esphome
