+++ esphome/components/coolix/coolix.cpp
@@ -27,8 +27,6 @@ const uint32_t COOLIX_FAN_MED = 0x5000;
 const uint32_t COOLIX_FAN_MAX = 0x3000;
 
 // Temperature
-const uint8_t COOLIX_TEMP_MIN = 17;  // Celsius
-const uint8_t COOLIX_TEMP_MAX = 30;  // Celsius
 const uint8_t COOLIX_TEMP_RANGE = COOLIX_TEMP_MAX - COOLIX_TEMP_MIN + 1;
 const uint8_t COOLIX_FAN_TEMP_CODE = 0b1110;  // Part of Fan Mode.
 const uint32_t COOLIX_TEMP_MASK = 0b11110000;
@@ -60,56 +58,7 @@ static const uint32_t FOOTER_SPACE_US = HEADER_SPACE_US;
 
 const uint16_t COOLIX_BITS = 24;
 
-climate::ClimateTraits CoolixClimate::traits() {
-  auto traits = climate::ClimateTraits();
-  traits.set_supports_current_temperature(this->sensor_ != nullptr);
-  traits.set_supports_auto_mode(true);
-  traits.set_supports_cool_mode(this->supports_cool_);
-  traits.set_supports_heat_mode(this->supports_heat_);
-  traits.set_supports_two_point_target_temperature(false);
-  traits.set_supports_away(false);
-  traits.set_visual_min_temperature(17);
-  traits.set_visual_max_temperature(30);
-  traits.set_visual_temperature_step(1);
-  return traits;
-}
-
-void CoolixClimate::setup() {
-  if (this->sensor_) {
-    this->sensor_->add_on_state_callback([this](float state) {
-      this->current_temperature = state;
-      // current temperature changed, publish state
-      this->publish_state();
-    });
-    this->current_temperature = this->sensor_->state;
-  } else
-    this->current_temperature = NAN;
-  // restore set points
-  auto restore = this->restore_state_();
-  if (restore.has_value()) {
-    restore->apply(this);
-  } else {
-    // restore from defaults
-    this->mode = climate::CLIMATE_MODE_OFF;
-    // initialize target temperature to some value so that it's not NAN
-    this->target_temperature = (uint8_t) roundf(clamp(this->current_temperature, COOLIX_TEMP_MIN, COOLIX_TEMP_MAX));
-  }
-  // never send nan as temperature. HA will disable the user to change the temperature.
-  if (isnan(this->target_temperature))
-    this->target_temperature = 24;
-}
-
-void CoolixClimate::control(const climate::ClimateCall &call) {
-  if (call.get_mode().has_value())
-    this->mode = *call.get_mode();
-  if (call.get_target_temperature().has_value())
-    this->target_temperature = *call.get_target_temperature();
-
-  this->transmit_state_();
-  this->publish_state();
-}
-
-void CoolixClimate::transmit_state_() {
+void CoolixClimate::transmit_state() {
   uint32_t remote_state;
 
   switch (this->mode) {
